moved swapeventsubscriber out of orderslice
